Fix last open host state
