+++ src/commands/viewSnippetInBrowser.ts
@@ -1,5 +1,5 @@
 import { Uri, env } from "vscode";
-import { Snippet } from "./types";
+import { Snippet } from "../types";
 
 export default function viewSnippet(snippet: Snippet) {
   return env.asExternalUri(Uri.parse(snippet.web_url)).then(env.openExternal);
